+++ frontend-react/modals/ModalsMobi/ModalStatusMobi/content.tsx
