+++ frontend-react/modals/modals.tsx
@@ -1,7 +1,8 @@
+'use client'
 import TestModalDesktop from '@/modals/ModalsDesktop/TestModalDesktop'
 import TestModalMobi from '@/modals/ModalsMobi/TestModalMobi'
-import ModalContractTerminatedMobi from './ModalsMobi/ModalContractTerminatedMobi'
-import ModalOrderAcceptedMobi from './ModalsMobi/ModalOrderAcceptedMobi'
+import ModalContractTerminatedMobi from './ModalsMobi/ModalStatusMobi/ModalContractTerminatedMobi'
+import ModalOrderAcceptedMobi from './ModalsMobi/ModalStatusMobi/ModalOrderAcceptedMobi'
 
 export const modals = {
     desktop: [{ id: 'test_desktop', content: <TestModalDesktop /> }],
@@ -9,12 +10,15 @@ export const modals = {
         { id: 'test_mobi', content: <TestModalMobi /> },
         {
             id: 'contract_terminated_mobi',
-            content: ({ onClose }: { onClose: () => void }) => <ModalContractTerminatedMobi closeModal={onClose} />,
+            content: ({ onClose }: { onClose: () => void }) => (
+                <ModalContractTerminatedMobi isOpen={true} onClose={onClose} />
+            ),
         },
-
         {
             id: 'order_accepted_mobi',
-            content: ({ onClose }: { onClose: () => void }) => <ModalOrderAcceptedMobi closeModal={onClose} />,
+            content: ({ onClose }: { onClose: () => void }) => (
+                <ModalOrderAcceptedMobi isOpen={true} onClose={onClose} />
+            ),
         },
     ],
 }
